Factor out file source file tests
