+++ tests/Application/AbstractFileSourceFileTest.php
@@ -2,38 +2,30 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Integration;
+namespace App\Tests\Application;
 
 use App\Entity\FileSource;
 use App\Services\Source\Store;
 use App\Tests\DataProvider\AddFileInvalidRequestDataProviderTrait;
 use App\Tests\DataProvider\TestConstants;
 use App\Tests\DataProvider\YamlFileInvalidRequestDataProviderTrait;
-use App\Tests\Services\EntityRemover;
 
-class AddReadDeleteFileTest extends AbstractIntegrationTest
+abstract class AbstractFileSourceFileTest extends AbstractApplicationTest
 {
     use AddFileInvalidRequestDataProviderTrait;
     use YamlFileInvalidRequestDataProviderTrait;
 
-    private FileSource $source;
-    private Store $store;
+    private FileSource $fileSource;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $entityRemover = self::getContainer()->get(EntityRemover::class);
-        if ($entityRemover instanceof EntityRemover) {
-            $entityRemover->removeAll();
-        }
-
         $store = self::getContainer()->get(Store::class);
         \assert($store instanceof Store);
-        $this->store = $store;
 
-        $this->source = new FileSource($this->authenticationConfiguration->authenticatedUserId, '');
-        $this->store->add($this->source);
+        $this->fileSource = new FileSource($this->authenticationConfiguration->authenticatedUserId, '');
+        $store->add($this->fileSource);
     }
 
     /**
@@ -47,8 +39,8 @@ class AddReadDeleteFileTest extends AbstractIntegrationTest
         array $expectedResponseData
     ): void {
         $response = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             $filename,
             $content
         );
@@ -66,8 +58,8 @@ class AddReadDeleteFileTest extends AbstractIntegrationTest
         array $expectedResponseData
     ): void {
         $response = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             $filename
         );
 
@@ -84,22 +76,44 @@ class AddReadDeleteFileTest extends AbstractIntegrationTest
         array $expectedResponseData
     ): void {
         $response = $this->applicationClient->makeRemoveFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             $filename
         );
 
         $this->responseAsserter->assertInvalidRequestJsonResponse($response, $expectedResponseData);
     }
 
-    public function testAddReadRemoveFileSuccess(): void
+    public function testRemoveFileNotFound(): void
+    {
+        $response = $this->applicationClient->makeRemoveFileRequest(
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
+            TestConstants::FILENAME
+        );
+
+        $this->responseAsserter->assertSuccessfulResponseWithNoBody($response);
+    }
+
+    public function testReadFileNotFound(): void
+    {
+        $response = $this->applicationClient->makeReadFileRequest(
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
+            TestConstants::FILENAME
+        );
+
+        $this->responseAsserter->assertNotFoundResponse($response);
+    }
+
+    public function testAddReadUpdateRemoveFileSuccess(): void
     {
         $initialContent = '- initial content';
         $updatedContent = '- updated content';
 
         $addResponse = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME,
             $initialContent
         );
@@ -107,16 +121,16 @@ class AddReadDeleteFileTest extends AbstractIntegrationTest
         $this->responseAsserter->assertSuccessfulResponseWithNoBody($addResponse);
 
         $initialReadResponse = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME
         );
 
         $this->responseAsserter->assertReadSourceSuccessResponse($initialReadResponse, $initialContent);
 
         $updateResponse = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME,
             $updatedContent
         );
@@ -124,24 +138,24 @@ class AddReadDeleteFileTest extends AbstractIntegrationTest
         $this->responseAsserter->assertSuccessfulResponseWithNoBody($updateResponse);
 
         $updatedReadResponse = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME
         );
 
         $this->responseAsserter->assertReadSourceSuccessResponse($updatedReadResponse, $updatedContent);
 
         $removeResponse = $this->applicationClient->makeRemoveFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME
         );
 
         $this->responseAsserter->assertSuccessfulResponseWithNoBody($removeResponse);
 
         $notFoundReadResponse = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->source->getId(),
+            $this->authenticationConfiguration->validToken,
+            $this->fileSource->getId(),
             TestConstants::FILENAME
         );
