+++ tests/Functional/Application/FileSourceFileTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Functional\Application;
+
+use App\Tests\Application\AbstractFileSourceFileTest;
+
+class FileSourceFileTest extends AbstractFileSourceFileTest
+{
+    use GetClientAdapterTrait;
+}
