--- tests/Functional/Controller/FileSourceFileControllerTest.php
@@ -1,199 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Controller;
-
-use App\Entity\FileSource;
-use App\Services\Source\Store;
-use App\Tests\DataProvider\AddFileInvalidRequestDataProviderTrait;
-use App\Tests\DataProvider\TestConstants;
-use App\Tests\DataProvider\YamlFileInvalidRequestDataProviderTrait;
-use League\Flysystem\FilesystemOperator;
-
-class FileSourceFileControllerTest extends AbstractSourceControllerTest
-{
-    use AddFileInvalidRequestDataProviderTrait;
-    use YamlFileInvalidRequestDataProviderTrait;
-
-    private FilesystemOperator $fileSourceStorage;
-
-    private FileSource $fileSource;
-    private string $sourceRelativePath;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $fileSourceStorage = self::getContainer()->get('file_source.storage');
-        \assert($fileSourceStorage instanceof FilesystemOperator);
-        $this->fileSourceStorage = $fileSourceStorage;
-
-        $this->fileSource = new FileSource(
-            $this->authenticationConfiguration->authenticatedUserId,
-            'file source label'
-        );
-        $this->sourceRelativePath = $this->fileSource->getDirectoryPath();
-
-        $store = self::getContainer()->get(Store::class);
-        \assert($store instanceof Store);
-        $store->add($this->fileSource);
-    }
-
-    /**
-     * @dataProvider addFileInvalidRequestDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testAddFileInvalidRequest(
-        string $filename,
-        string $content,
-        array $expectedResponseData
-    ): void {
-        $response = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename,
-            $content
-        );
-
-        $this->responseAsserter->assertInvalidRequestJsonResponse($response, $expectedResponseData);
-    }
-
-    public function testAddFileSuccess(): void
-    {
-        $filename = TestConstants::FILENAME;
-        $content = '- file content';
-        $fileRelativePath = $this->sourceRelativePath . '/' . $filename;
-
-        self::assertFalse($this->fileSourceStorage->directoryExists($this->sourceRelativePath));
-        self::assertFalse($this->fileSourceStorage->fileExists($fileRelativePath));
-
-        $response = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename,
-            $content
-        );
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertTrue($this->fileSourceStorage->directoryExists($this->sourceRelativePath));
-        self::assertTrue($this->fileSourceStorage->fileExists($fileRelativePath));
-        self::assertSame($content, $this->fileSourceStorage->read($fileRelativePath));
-    }
-
-    public function testUpdateFileSuccess(): void
-    {
-        $filename = TestConstants::FILENAME;
-        $initialContent = '- initial content';
-        $updatedContent = '- updated content';
-        $fileRelativePath = $this->sourceRelativePath . '/' . $filename;
-
-        $this->fileSourceStorage->write($fileRelativePath, $initialContent);
-
-        $response = $this->applicationClient->makeAddFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename,
-            $updatedContent
-        );
-
-        $this->responseAsserter->assertSuccessfulResponseWithNoBody($response);
-        self::assertTrue($this->fileSourceStorage->directoryExists($this->sourceRelativePath));
-        self::assertTrue($this->fileSourceStorage->fileExists($fileRelativePath));
-        self::assertSame($updatedContent, $this->fileSourceStorage->read($fileRelativePath));
-    }
-
-    /**
-     * @dataProvider yamlFileInvalidRequestDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testRemoveFileInvalidRequest(
-        string $filename,
-        array $expectedResponseData
-    ): void {
-        $response = $this->applicationClient->makeRemoveFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename
-        );
-
-        $this->responseAsserter->assertInvalidRequestJsonResponse($response, $expectedResponseData);
-    }
-
-    public function testRemoveFileSuccess(): void
-    {
-        $filename = TestConstants::FILENAME;
-        $content = '- file content';
-        $fileRelativePath = $this->sourceRelativePath . '/' . $filename;
-
-        $this->fileSourceStorage->write($fileRelativePath, $content);
-
-        $response = $this->applicationClient->makeRemoveFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename
-        );
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertFalse($this->fileSourceStorage->fileExists($fileRelativePath));
-    }
-
-    public function testRemoveFileNotFound(): void
-    {
-        $response = $this->applicationClient->makeRemoveFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            TestConstants::FILENAME
-        );
-
-        $this->responseAsserter->assertSuccessfulResponseWithNoBody($response);
-    }
-
-    /**
-     * @dataProvider yamlFileInvalidRequestDataProvider
-     *
-     * @param array<mixed> $expectedResponseData
-     */
-    public function testReadFileInvalidRequest(
-        string $filename,
-        array $expectedResponseData
-    ): void {
-        $response = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename
-        );
-
-        $this->responseAsserter->assertInvalidRequestJsonResponse($response, $expectedResponseData);
-    }
-
-    public function testReadFileNotFound(): void
-    {
-        $response = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            TestConstants::FILENAME
-        );
-
-        $this->responseAsserter->assertNotFoundResponse($response);
-    }
-
-    public function testReadFileSuccess(): void
-    {
-        $filename = TestConstants::FILENAME;
-        $content = '- file content';
-        $fileRelativePath = $this->sourceRelativePath . '/' . $filename;
-
-        $this->fileSourceStorage->write($fileRelativePath, $content);
-
-        $response = $this->applicationClient->makeReadFileRequest(
-            $this->validToken,
-            $this->fileSource->getId(),
-            $filename
-        );
-
-        $this->responseAsserter->assertReadSourceSuccessResponse($response, $content);
-    }
-}
