+++ tests/Integration/FileSourceFileTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Integration;
+
+use App\Tests\Application\AbstractFileSourceFileTest;
+
+class FileSourceFileTest extends AbstractFileSourceFileTest
+{
+    use GetClientAdapterTrait;
+}
